Detokenize data used for string folding
The older version stores the tokens with whatever catcode they've been given
by the format: that will produce the wrong output. By detokenizing at point of
reading the folding code itself can be keep simple (check-in later today, I 
hope). Note that there is a XeTeX bug with chars beyond 0xFFFF which can be
handled at least to some extent if everything is detokenized (probably it ends
up applying twice, once to the 'search' and once to the 'replace' chars, so
cancels out).


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5152 de43f980-851b-0410-b2f7-c40aca1f87e0