add \tl_set_from_file(_x):Nnn and variants to l3candidates
I think we discussed this would be a good idea; the
implementation is a fairly literal port from Heiko's catchfile.

I'm not 100% sure that the "_x" version is needed,
(I've never used it) but it can't hurt to add for now.

I assume we do want to use the \tl_ prefix here and not \file_.

Sorry, too late to add test files tonight :(

git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5160 de43f980-851b-0410-b2f7-c40aca1f87e0