Tidy up \tl_(g)set_from_file(_x):Nnn a bit
In the docs, this is mainly editorial: stick to the style we generally have
in our stuff.

In the code, as well as some editorial stuff and a few comments on what is
going on, there are a few changes. I've avoided needing anything global and
have used the "@@" convention. The code is also now in the right place in the
file.