New Unicode-baed case changing functions
As discussed on the team list a while ago, having 'proper' case changing
functions is one of the things required to move away from the \lowercase/
\uppercase primitives being used 'directly'.

The implemenation is x-type expandable, as for \str_fold_case:n. However, as
the input here is tokenized the code is based partly on \__tl_act:NNNnn.
Naming-wise, I am not sure if \tl_upper_case:n, etc., is 'right' but it is
not I think too bad and similar to \str_fold_case:n.

The core stuff here follows Unicode more-or-less completely. Context-dependent
stuff is more tricky, as from our point of view 'look behind' matches are
awkward. Note that the Unicode docs tend to be written in very general terms
but in reality the number of possible sensible matches is quite small. One
think that may need more work is the 'final sigma' business, as I've gone for
a relatively simple approach that perhaps leaves more cases sub-optimal but
gains a decent amount of performance.

As with the string function, the approach here only gives 'full' results using
a UTF-8 engine.

One obvious question is whether the 'skip brace group' approach is right: it's
similar to BibTeX so is at least familar to many people.

If this approach/naming looks OK I suggest we deprecate
\tl_expandable_(lower|upper)case:n and look to remove end-2015.