Add Dutch exception to case changing
In Dutch, "ij" is capitalised as "IJ" at the start of words. As this is a
simple ASCII case the change will work with pdfTeX as well as XeTeX/LuaTeX:
I'm assuming that if the user has two chars in the input they want two in
the output (there is a single Unicode codepoint for the combination in
all three cases). 


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5164 de43f980-851b-0410-b2f7-c40aca1f87e0