Add Dutch exception to case changing
In Dutch, "ij" is capitalised as "IJ" at the start of words. As this is a
simple ASCII case the change will work with pdfTeX as well as XeTeX/LuaTeX:
I'm assuming that if the user has two chars in the input they want two in
the output (there is a single Unicode codepoint for the combination in
all three cases).