Various improvements to \tl_(g)set_from_file:Nn (closes #177)
Most of these are taken directly from suggestiosn by Qing Lee. I've used a
slightly 'cleaner' approach for the expansion-out-of-a-group chain. The
one place I've not changed is where \tl_set:Nx is used in storing the grabbed
file. This is needed for the same reason we use \edef internally for \tl_set:Nn:
it allows for `#` tokens.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5171 de43f980-851b-0410-b2f7-c40aca1f87e0