Add \box_resize_to_wd_and_ht:Nnn
As discussed on LaTeX-L, this goes with \box_resize_to_ht:Nn. As well as
adding the function, the test file here is now extended to cover this case
with the demos actually using depth (the old ones did not), and with a few
editorial changes (e.g. use \l_tmpa_box rather than create a new one, drop
space between numbers and units as this is probably 'better' for TeX despite
the fact units are typeset with spaces here!).