Remove unnecessary parentheses in some fp expressions
We include unnecessary braces around N-type arguments passed
to n-type functions, as this costs nothing.  The situation is
different for l3fp, as parentheses come with a non-negligible
overhead.  No need for them in expressions like "a*b+c*d".
We could also remove them from "sin(\l_..._fp)", but that
looks a bit offensive to me.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5193 de43f980-851b-0410-b2f7-c40aca1f87e0