Remove unnecessary parentheses in some fp expressions
We include unnecessary braces around N-type arguments passed
to n-type functions, as this costs nothing.  The situation is
different for l3fp, as parentheses come with a non-negligible
overhead.  No need for them in expressions like "a*b+c*d".
We could also remove them from "sin(\l_..._fp)", but that
looks a bit offensive to me.