Move \<thing>_item:Nn and :nn functions to stable
This commit means that

 - \clist_item:nn
 - \clist_item:nn
 - \seq_item:Nn
 - \tl_item:Nn
 - \tl_item:nn

are now all stable and adds

 - \prop_item:Nn

At the same time, \prop_get:Nn is deprecated. As the latter is in use 'in the
wild' in various packages, it is added to the standard deprecation cycle (one
full year plus lead-in), so will be removed at the end of 2015.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5214 de43f980-851b-0410-b2f7-c40aca1f87e0