Change precedence of juxtaposition in l3fp expressions
This is a breaking change!  But it should not affect too much code.

As discussed recently on LaTeX-L, the previous precedence of
juxtaposition was not a good choice, as it led to surprises such as
2pi^2 giving (2*pi)^2 instead of the expected 2*pi^2.  Now
juxtaposition simply causes l3fp to insert *, thus it has the same
precedence as * and /.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5223 de43f980-851b-0410-b2f7-c40aca1f87e0