Hard-code name for l3str-expl.dtx
As \GetIdInfo picks up names from the SVN data, we end up here with a file
name "l3str-expl". However, the DocStrip code unpacks just as "l3str"
(see earlier commits and new l3str module in l3kernel for reasons). Thus the
name here has to be forced so LaTeX2e does not issue a warning when loading
the experimental part of the code.