Use delimited argument instead of str_if_eq in n-type quark tests
Two advantages.  Accuracy: string comparison is not able to
distinguish \q_nil from the result of \tl_to_str:n { \q_nil }.
Speed in LuaTeX: since string comparison happens on the Lua side,
there is an overhead (escaping the string for Lua in particular).
Concretely, the new code is 1.2x to 2x slower in pdfTeX but
about 3x faster in luaTeX.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5251 de43f980-851b-0410-b2f7-c40aca1f87e0