Allow for ConTeXt's handling of some Unicode chars when reading case data
Lua callbacks can be used when reading files: this is done by ConTeXt but
in principle could happen with any LuaTeX code. This can be used to rewrite
chars, which in the case this checking covers means replacing compat chars
with their potentially multi-char equivalent. That's a problem for us as
it can mean that one char turns into two at point of use, which if they are
unbraced can mean that the mapping code used in case changes gets the wrong
number of args and things go wrong.

The fix here is a bit involved: the aim is to end up making as few changes
as possible to the data file, so there are more in the set-up file to enable
this. UnicodeData.txt is read before the string data and all compat chars that
lead to a change in the number of chars are noted. Whenever those cases are
then written to the data file braces are added. The result is that reading back
will always work correctly, Lua callback or not, while in most cases there
are no braces added.