Make verbatim commands ignore spaces before leading delimiter
Previously, with '\foo' of signature 'm s v', the call

    \foo{}*
    ++

would fail (the space after '*' cannot be used as a delimiter),
while the same call without the star would succeed, since the
space would be removed when looking for the optional star.  Now
for consistency the 'v'-type argument ignores spaces when
looking for its starting delimiter.  This will be necessary for

    \begin{macro}[...]
      |\...,\...|

to work.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5301 de43f980-851b-0410-b2f7-c40aca1f87e0