Add dependeny on xparse for l3kernel doc building
Bruno's recent changes mean that an up-to-date xparse is needed to get
"build doc" to complete. That's a general thing: as l3doc uses xparse, there
shoudl be a dependency in the scripts.

(This is I guess a bit circular, but we know that all of the l3 stuff is
essentially one product split into parts for convenience, so hopefully it is
acceptable.)


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5305 de43f980-851b-0410-b2f7-c40aca1f87e0