Add dependeny on xparse for l3kernel doc building
Bruno's recent changes mean that an up-to-date xparse is needed to get
"build doc" to complete. That's a general thing: as l3doc uses xparse, there
shoudl be a dependency in the scripts.

(This is I guess a bit circular, but we know that all of the l3 stuff is
essentially one product split into parts for convenience, so hopefully it is
acceptable.)