Avoid multiply-defined label by removing one of the docs
The function "\or:" is documented with "\if_case:w" as it should,
and the documentation with "\if_false:w", "\if_true:w", "\else:",
"\fi:" didn't really make sense.

Now there should only remain two multiply-defined labels:
"doc//function/+" and "doc//function/-" (binary and unary).
I don't know what to do: it would make sense for them to appear as
"+ (unary)" versus "+ (binary)" in the index, but that's not easy.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5351 de43f980-851b-0410-b2f7-c40aca1f87e0