Extend 'version interlock' to some of l3experimental
This addresses the CTAN worry about version data in uploads and also is aimed 
at avoiding any weird version clashes. I have not added the version data to
every dependent package, but hopefully to enough that the idea works. It also
raises a few more issues: to be addressed separately.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5374 de43f980-851b-0410-b2f7-c40aca1f87e0