Normalise LuaTeX's \csname\endcsname bug
At present, LuaTeX prints "\ycsnam\yendcsnam" rather than "\csname\endcsname".
That could be covered by a LuaTeX-specific test file, but the change DEK made
to printing a space after "\csname\endcsname" means that depending on the TeX
version there's also a space variation. We normalise ou the latter for
"\csname\endcsname" but obviously that misses the LuaTeX case. The change here
first deals with the LuaTeX bug so that our normalisation for the DEK change
works for all engines. We then don't need a separate test file for LuaTeX
here.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5392 de43f980-851b-0410-b2f7-c40aca1f87e0