l3build: Address the \askforoverwrite issue
When \askforoverwrite is true, TeX will ask for confirmation of file
creation if the file exists. It's difficult to arrange that no file is
generated twice, so despite the fact that unpacking is done in a 'safe'
place we could still get the warning. The solution is to create a file
containing lots of "y\n", and then feed that to TeX as part of the unpack
process.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5442 de43f980-851b-0410-b2f7-c40aca1f87e0