Fix handling of Unicode case changing exceptions
The old version here only covered exceptions where UnicodeData.txt has a
mapping but not where the only data is in SpecialCasing.txt.

At the same time, some refactoring internally makes sure that \tl_rescan:nn
applies to all cases required as some other high-position data was also
wrong.

The SVN field on the .def file is not set as this is going to see futher
revision in a second commit.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5464 de43f980-851b-0410-b2f7-c40aca1f87e0