Drop use of 'lualatexquotejobnmae.lua'
This file is used to add jobname quoting to LuaTeX, which happens with
web2c versions of TeX, pdfTeX, XeTeX 'out of the box'. As reasoned
for LaTeX2e, this is really a 'distro' think (we can't control the
behaviour in engines other than LuaTeX). As such, this step should be
left to distro people.

Note this change only affects format mode.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5490 de43f980-851b-0410-b2f7-c40aca1f87e0