Update case change functions to use textcase-like approach to escaping
Rather than the BibTeX approach, with this update escaping requires
a control sequence argument be used. There are no pre-defined functions
for general escape although one could be provided: with LaTeX2e the same
set of sensible escapes is used as for textcase.

This is the first change after recent discussion on LaTeX-L: others
will follow in due course over the rest of today/tomorrow.