source files from testfiledir on check or save
Test files may still be placed in testfiledir, even when packed. Keeping
them there adds the benefit that they will only be considered with the
check and save families of commands.

This introduces a new parameter to the bundleunpack function, without
changing its behaviour. It can be used to specify directories to search
for the sourcefiles and defaults to the previously used local directory.