Add \str_lower_case:n and \str_upper_case:n
These are needed for the case in the documentation example:
on-the-fly control sequence construction such as \textAlpha (see
LaTeX2e package "chemgreek").

To make these functions available and expandable and to share data with
the other case commands, some rearrangements are made. The Unicode data
is now stored in two parts: the simple 1-1 mappings and more complex
cases. This has a slight performance implication for 'text' case changing
and requires detokenization of the results for \str_... commands. There
is therefore a knock-on on a number of the existing functions but only
at the implementation level.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5544 de43f980-851b-0410-b2f7-c40aca1f87e0