l3build: Run Biber if required
This merges (more or less) pull request #207.

As noted there, currently only Biber is covered from a biblatex point of
view: BibTeX is rather tricker as it will need either a swtich or some
code to parse the .aux file.

Another question in this area is how to deal with .bib files: still needs
thinking about for full support for bibliographies.


git-svn-id: http://www.latex-project.org/svnroot/experimental/trunk@5557 de43f980-851b-0410-b2f7-c40aca1f87e0