Update to TL'15 for tests
There are only a small number of differences between TL'14 and TL'15,
most down to the new allocator in LaTeX2e. Loading etex is dropped here
meaning that there are a few adjustments to register use. XeTeX now
ends up using one fewer register than pdfTeX: this happens as the
driver needs a register for pdfTeX but not for XeTeX, but that was
previously masked by the format needing one more register for XeTeX
(interclasschar allocator): the latter is now in the extended pool.

There are a couple of new LuaTeX failures, to be addressed by further
updates to l3build in terms of normalisation.

Changes here impact only on the LaTeX3 repo itself: no knock-on for
LaTeX2e or others.