l3build: Basic BibTeX support (see #208)
This runs BibTeX if required but there is not yet
support for extracted .bst or .bib files, etc. The idea of using 
findstr on Windows as approx. grep on Unix is from dffisscher
(thanks).

There are some issues this points up with local search, etc.,
which will need to be addressed (not just for BibTeX): needs a
little testing.