l3build: Remove need for --output-directory for typesetting (see #213)
This fixes most of #213 but there are still a few niggles outstanding.
Broadly, this removes the need to mess about with the typesetting
location, so also any interference from auxiliary files 'here'. Some
tidying up of other parts of the code will follow, including addressing
the secondary tools part of the issue.