l3build: Modularise typsetting
This makes it possible to customise typesetting beyond the simple
variable-based approach. Thus we can have a relatively simple build
process suitable for common cases but still allow users with other needs
to harness l3build. To complete this we need to address BibTeX .bib and .bst
files: that should happen over the next day or so.