Extend some documentation
Doing expansion of math mode material is possible but it seems at odds with
the idea that math mode is supposed to be skipped by case changing. As such,
it seems preferable to simpy leave this alone and to document that the
closing token has to be in the text itself. It's very much an edge case
anyway: in real 'text' the only thing that should be being expanded is e.g.
a saved name.