Overhaul setting of parshape
The new approach sets the paragraph shape purely from the galley
data without any attempt to recover things from the primitive. That
means that we always need to do margins/measure/cutouts but makes the
overall flow clearer. 

The set up here works even if the \par token occurs at the end of
a group and a cutout is active.