Fix iow_indent in first line of iow_wrap (see #215)
The difficulty is that \iow_indent:n needs to translate to " marker " with
spaces in order to separate the marker from surrounding words.  When
\iow_indent:n appears at the start of a line, the leading space makes an empty
word, which is placed together with a space in the line being built.  The line
is then non-empty, and using \iow_indent:n in a non-empty line makes the
indent start on the next line.  This explains both the single space and the
absence of indent in

    \iow_wrap:nnnN { \iow_indent:n { ... } } { } { } \iow_term:n

I had included code to avoid this problem when \iow_indent:n followed \\, but
not at the very beginning.  Now the string always starts with (essentially) \\