Add all LuaTeX/Umath primitives to l3names
To provide a consistent interface for accessing primitives we need
reasonably-comprehensive coverage of primitives. Here, that is done
for LuaTeX with knock-on effects elsewhere.

All primitives added by LuaTeX, including those from Omega and Aleph
that it provides, are now included.

All pdfTeX primitives that are also available in LuaTeX are included.

All Umath primitives are includes, covering those shared with XeTeX
and all others.

To think about/check:

 - pdfTeX primitives are divided into those related to and not
   related to PDF mode: that's not all easy so there may be a
   few mistakes

 - Umath primtives have all been called \umath_...: OK? If
   so, do the names make sense for example for \Umathcode =>
   \umath_code:D? The "math" might be best retained here in
   the name but which cases should that include?

 - Umath primitives may be \luatexUmath... only in LaTeX2e
   depending on format date. I've enabled the raw names to
   avoid a lot of repetitive code: OK?

 - XeTeX/LuaTeX/pdfTeX similarities are hoepfully merged but
   this leave \pdfmap(file|line) a bit odd: see comments and
   test.