Make key properties work in additive fashion
As dicussed on LaTeX-L, this enables the removal of key defintions using
the new .undefine: property. The prop used to store key data is now always
defined only locally, which here makes sense. The result is that there is
less noise in the log from keys than there was.

Altering .value_required: and .value_forbidden will be a separate step:
later today or tomorrow.