Pass data to messages more cleanly, simplify \__msg_show_variable:Nnn
Instead of passing a csname to the show-clist, show-seq, show-prop
messages, pass as a separate argument (equal to '' or '?') the
information of whether the clist/seq/prop is empty or not.  To get the
height, depth and width of a coffin into a message, use custom code
instead of \__msg_show_variable:Nnn.  As a result, some code that was
needed there to deal with coffins can be removed.
use custom code for coffins