Make \char_set_catcode add their arg to \dospecials (fixes #45)
In package mode, \char_set_catcode:nn (hence all functions
built on it) adds the character to \dospecials (in TeX/LaTeX)
and \@sanitize (in LaTeX), unless the catcode is 11 or 12 in
which case it is removed instead.  There is some code to
avoid ending up with several copies of the same character.

I'm not sure what the right way to test this code is.
Right now my test logs the length of \dospecials and
\@sanitize before/after a \char_set_catcode:nn.