Drop code that keeps \l_char_active_seq up to date (see #45)
The sequence \l_char_active_seq is meant to contain active
characters, but the active ^^L is \outer in LaTeX2e and plain
TeX, and I don't know whether to special case that or do
something else.

This same commit fixes another problem with \outer commands:
in \dospecials (plain&2e), \@sanitize (LaTeX2e), and
\l_char_special_seq, characters are stored as single-character
control sequences, for instance # as \# etc.  The issue is that
\+ is outer in plain and 2e, so it cannot be manipulated.
Thus, the new code here only constructs \csname <char>\endcsname
if that command should be stored or removed from \dospecials
etc.