Provide control of how round rounds ties (fixes #222)
In a floating point expression "round(x, n)" rounds
the floating point x to n places (a positive or negative
integer), and ties are rounded such that the last digit
that remains is even.  Now "round(x, n, t)" provides
control on the rounding of ties: if t=0 or t=-0, ties
are rounded towards zero; if t>0 they are rounded up;
if t<0 they are rounded down; if t=nan they are rounded
to an even number as is done by default.

An alternative would be to provide functions with their
own name: roundtiesup, roundtiesdown, roundtieszero, or
roundceil, roundfloor, roundtrunc, or something else.
I'm happy either way but just wanted to be done with it.