Rewrite \tl_rescan/\tl_(g)set_rescan (fixes #154, fixes #48)
I decided to distinguish single-line strings from multi-line
strings when rescanning.  I see the multi-line case as being
useful for running code that was collected from a file using
e.g. \ior_get_str:NN, thus \scantokens is used directly, not
doing any trick to try and preserve spaces, as TeX would not
preserve them when reading from a file.  For the single-line
case, tokens are rescanned as if they appeared in the middle
of a line.  Leading and trailing spaces and/or tabs are kept.
The two cases are distinguished automatically.  Is providing
an override necessary?