In char_set_catcode, do a csname expansion in a group
To avoid letting the single-character control sequence
(built for updating \@sanitize) equal to relax if it
was undefined, perform expansion from within a group.
This uses \exp_args:NNNx, thus added to l3expan.dtx.
Thanks to Qing Lee.  Note that in practice the control
sequences involved are \{ and \}, which are not
undefined.