Parsing big-integer expressions (addition/subtraction) (see #251)
The internals of the package are completely rewritten to allow for
big integer expressions, with a plan to support (+-*/"'`) in the
(hopefully near) future.  Right now only addition and subtraction
have been implemented.  Multiplication and division are missing,
as is \bigint_compare:nTF (nNnTF is ok), hence the much reduced
test file.