Fix big integer addition (it added together mismatched parts of numbers)
It is complicated to describe the bug. The code splits both numbers in 
blocks of 8 digits, organized into two binary trees, then corresponding 
blocks of digits are brought close together in the input stream, and 
summed.  Schematically, a 64-digit number with blocks ABCDEFGH becomes a 
tree {{AB}{CD}}{{EF}{GH}}.  A 48-digit number IJKLMN used to become 
{IJ}{{KL}{MN}}, and ABCDEFGH + IJKLMN thus did the following: the 
addition was split into {AB}{CD}+IJ and {EF}{GH}+{KL}{MN}, then the 
first calculation was split as {AB}+I and {CD}+J.  This is obviously 
wrong.  Now essentially there is an extra pair of braces around IJ so 
that {AB}{CD}+{IJ} becomes {AB}+0 and {CD}+{IJ}.

I have an implementation of Karatsuba's multiplication algorithm, but
so far it only covers cases where the number of digits is a power of 2.
Expect a commit tomorrow when I'm done.