Remove avoidable use of \tex_lowercase:D
In anticipation of addressing the status of \tl_to_lowercase:n,
the first stage is to tidy up internal use of case changing. Here,
places where an alternative for use of the 'raw' primitive are
addressed. (One use is left in l3str-format-new.dtx but relates to
\fp_to_scientific:n, which itself needs adjustment for other reasons.)

The use of \tl_to_lowercase:n and provision of new interfaces for
generating 'odd' characters is outstanding. Note there is one change
from \tex_lowercase:D to \tl_to_lowercase:n here which will later
change again.

Note that there are no uses of \tex_uppercase:D in the code base.