Intergrate l3str into the kernel
This commit adds all of the 'experimental' part of l3str to expl3.
The primary driver here is str variables, but it seemed best to
take everything.

There are a few adjustments to other places to accommodate this.
Most notably, this now means loading l3str alone gives an error,
but only a couple of packages outside of the repo do this. Also
notice loss of some (old) deprecated functions: as these are in
l3experimental, no notice is needed.

Further adjustment is likely needed yet. In particular, I will
probably split up the l3str part inside l3experimental since it
no longer creates a file "l3str.sty"!