Simplify getting char with different catcode in xparse
The only goal was to get a character that looks the same as
the end-delimiter of a D-type arg in the terminal but is not
matched by the grabber.  Instead of using catcode 11 or 7
depending on the catcode of the delimiter, use catcodes 11
or 12, the latter being much easier to produce.