l3build: Add "setversion" target
This will allow auto-editing of files to step version:
needed by us but also useful to others. At present this is
not documented: to do.

Note that as standard there is no attempt to edit: the user has to
define the appropriate function to do the work. Also note that
the file is read one line at a time: this avoids the need to match
for os_newline as it allows normal Lua pattern use for start/end of
line.