Add a font to the format (closes #231)
This is 'just enough' code to get some kind of output in format
mode. Most notably, there are no hyphenation patterns nor any output
routine. The latter is therefore covered by TeX's default OR which
at least gets some typesetting done.

Note that the 'classical' font is loaded even with XeTeX/LuaTeX at
present. For the latter we'd need a full font loader to use LM in
OpenType format, and how to handle that is still complex (will likely
need some glue between the format and luaoftload). Thus we stick at
present to soemthing that will at least work in isolation. Also
notice that the font is not loaded during format building so there
is no impact on sandboxing.