Ensure \Ucharcat emulation works for LuaTeX in IniTeX mode
This is all rather temporary as ideally an allocator should be
available. Probably this part needs writing in two entirely
separate blocks, one for package mode and one for IniTeX mode.
It is also probably worth considering a separate .lua file for
the Lua code: there will be more to add over time.