Update engine requirements to LuaTeX 0.70
The code to emulate \Ucharcat needs the correct Lua support:
this perhaps could be avoided but it would be awkward. Update
the docs as a result, leaving things general in terms of primitive
requirements as they are now somewhat more complex to explain!

Note that at some point XeTeX will need to be stepped to v0.99992
as that provides \Ucharcat.