Simplify \char_set_catcode:nn
The idea of 'auto-updating' the list of special chars is
attractive but problematic: for plain and 2e outer tokens
cause issues but more widely we are at the wrong level to
tell if the change is really document-wide or a local
'trick'.

Upcoming is also a change to how \l_char_active_seq works.