Add some ideas on input encoding
The idea here is that in contrast to LaTeX2e we should probably
start from an assumption of UTF-8 input in all cases. That will
make loading hyphenation patterns rather easier when we come to
it. So for the 8-bit (like) engines we can include the appropriate
conversion code in the format itself.

The upTeX situation is complicated as this might all go horribly
wrong when \disablecjktokens is not active: need to see some real
use cases to know.

Whether having a fixed input/font encoding is workable is very much
an open question but at least for testing this seems a reasonable
approach (if only to possibly rule things out).