Fix bug: r-type args were inconsistently optional/mandatory
This commit fixes the bug mentionned in my last commit.  It
has the side-effect of making a trailing r-type argument allowed.
One could argue that if an r-type arg is missing xparse would
wrongly have braced the next token.  But since a missing r-type
arg is an error, I think it's ok to have further problems.