Use Unicode data from TeX codes where appropriate
This new approach requires no 'pre-processed' data, only
the raw Unicode files and what is already loaded by
XeTeX/LuaTeX. For these engines, it is reasonable to assume a
fixed set of catcodes, etc., as the Unicode mappings are clear.

Note that pdfTeX, etc., stay 7-bit here but could be extended
to 8-bit if we really want. (Probably that is 'safe' for format
mode but not for package mode, as in the latter encodings can't
readily be predicted.)