New LuaTeX v0.87 primitives
These are the ones listed in the release notes: \(h|t|v)pack are
not currently in the manual! There are also other new primitives
listed by

    \directlua{print(tex.extraprimitives())}

but as they are not docmented and we are told

> there are some experimental new primitives but these will be
> described when they are stable (currently being tested)

I've missed those out. Even then, some of the additions here may
yet be reversed later!