l3build: Drop space-stripping at start of lines
The rational for stripping spaces is working with LuaTeX, but
if turns out it is best to leave line lengths unaltered outside of
the LuaTeX-specific code. (Upcoming commits will address LuaTeX.)
As such, this commit moves the .tlg files closer to the raw .log
file output.

This is going to require many .tlg files are rebuilt: LaTeX2e
checkin coming up for this.