New approach to UTF-8, accents and letter-like case changing
This new approach works with the T1-subset of UTF-8 even with
pdfTeX. The accent and letter-like code is also revised to be
consistent across engines.

As a result, pdfTeX input is assumed to be UTF-8. Currently this
may cause errors if the last char of the input looks like the first
octet of a UTF-8 sequence. That can be addressed with an extra
set of tests, but this is a 'policy' question as much as anything.

At present there are a few outstandings:
- Treatment of ß as a start-of-work char with pdfTeX
- Treatment of letter-like control sequences in mixed case