Add a simple font loader for LuaTeX in format mode
This allows us to load OpenType fonts 'out of the box', at least
if they are in the texmf tree. There's no font shaping here so its
only meant for European glyphs and even then limited. However, for
testing this should be handy.

Note that the code itself here is taken from Speeddata Publisher
with permission (see notes in .dtx). We can re-license but at
present I've gone with loading rather than copying so things are
clear. We are unlikely to use such a simple loader long-term in any
case. (The file fonts.lua from Speeddata Publisher shows how to do
basic shaping as needed for European support.)

Probably the next thing to do is deal with TeX ligatures and the
like, which will need a bit of work. That's all quite doable, as is
I think handling system fonts if we want to.

The big picture idea here is that we can now test Unicode stuff,
including hyphenation, in IniTeX mode.

Note that the format-building process is still a bit 'rough': you
need to put l3format.lua and fontloader.lua where LuaTeX can find
them.