Revise driver code
The structure here is new: essentially split into separate parts rather
than trying to DocStrip everything in one go. This makes the logic
for each driver easier to follow, and reflects experience with LaTeX2e.

The dvisvgm driver is now fully-functional and can do clipping.

Some revisions are included to the other drivers, notably a better
approach to colour for (x)dvipdfmx and avoidance of one definition
in dips.