Re-write of \keyval_parse:NNn
This is the first phase of dealing with #294. Whilst the bigger issue
is the higher-level code (which was largely written ad hoc to solve
the issue rather than to work efficiently), the code at the low level
has evolved over time without a proper check over.

Here, the idea is to take the basic keyval implementation (which is the
fastest the job can be done in TeX) and to extend it for our
requirements. In particular, there is a need to address some brace
stripping oddities in keyval that can't be altered there but can for
us. At the same time, our documented space stripping removes all
spaces not just one at each end, which requires a little more work.

Making the code 'babel safe' costs some time, but hopefully won't
be necessary at some stage in the future.

Testing with 10k runs of a setting for 4 keys and 2 blank entries,
and taking keyval as a baseline, the old version is about 6x while
the new version is about 2.3x. A non-babel safe version gets down to
about 2x.