Fix handling of cutouts if \par is in a group
In a group, the removal of items from the tracking seq works
but once these are empty there was no trigger to reset
\parshape. As the common case is no shape to do, it is
desirable to avoid setting \parshape every time. The approach
now is to note that we can tell if we have reached the top
level of the galley (at which point we are OK), and until we
have to set the shape for every par after a cutout starts.