Drop most of nopar in l3kernel (bar basics/expan/tl)
Other than the three 'tricky' modules this is largely a
mechanical excercise. The three outstanding modules need
more thought as the nature of the commands produces is
tricker: to be done one at a time.

There are a few edge case questions still open re.
what is an 'document' command and how they should be
set up.