Reduce use of nopar in l3basics
This module obviously needs a reasonable number of
occurences of nopar. Removal here leaves things like
the definition of \l__exp_internal_tl alone. The nature
of \prg_do_nothing: is also unchanged as this may be relied
on elsewhere (to check).

The very early definitions, of \cs_set_protected:Npn, etc.,
have been adjusted not only for the reduction of nopar but
also as mixing \tex_long:D with e.g. \cs_set_nopar:Npn
seems sub-optimal. They now expand to the primitives directly.
There is no change in the behaviour nor number of expansion
steps as \cs_set_nopar:Npn is \let to \tex_def:D in any case:
the change is to make the fact that this is 'low level' and
requires TeX knowledge clear.