l3doc: add keys "documented-as" (macro) and "label" (function)
See next commit for uses.  The documented-as key of the macro
environment is meant for cases which define variants rather
than the original function: this messes up the link to the
documentation.  The label key is used to avoid duplicates
labels when a function is documented multiple times (as done
in l3expan).