l3doc: Keep track of TF as an argument of auxiliaries (fixes #116)
Earlier, TF was appended to macro/function names according to
the value of some local boolean.  To typeset predicates this
boolean was locally set to false, but since functions are
typeset in a tabular, the setting did not last long enough,
leading to absurd displays like \bool_if_p:cTF.  Using a
global variable would be messy so I decided to carry the
boolean as an argument of every auxiliary that needs the info.