Add command to print help about star and TF notations
In l3doc, expandable functions are marked with a star, restricted
expandable with a hollow star and TF conditionals also have a
special notation.  When typesetting source3, every star and TF were
links to the explanation, but for other packages these links failed.
Now third-party packages can use \CodedocExplain, or the more specific
\CodedocExplainEXP, REXP, or TF, to have the explanation.  As an
example I've put that in l3sort in a footnote.