Add random numbers to l3fp, allow 0-argument functions
This commits adds rand() producing floating points in [0,1)
and randint(a,b) producing integers in [a,b] where a and b
must be integers less than 10^16 in magnitude (in principle
we could have allowed +10^16 and -10^16 too but it was easier
to code that way), and a can be omitted, defaulting to 1.

I reworked parsing a bit to allow the 0-argument rand() so
empty arguments (for instance due to extra commas) are now
ignored completely.  Thus max(,,,a,b,,,,) is equivalent to
max(a,b).  Drawback: round(a,,1) is equivalent to round(a,1),
where a user might have expected it to be equivalent to
round(a,0,1) namely "round to integer with ties rounded
toward +inf".