Use OT1 CMR for all engines
Following the 2017/01/01 LaTeX2e kernel update, the font used
'out of the box' is engine-dependent. That leads to differences
we are not interested in (probably).

For the present, force OT1/CMR at the level of the local .cfg
file. It may be worth adding this to the standard l3build set up
(when LaTeX is in use): to discuss.

Longer-term it may be sensible to add more .tlg files rather
than go this way, but that should be done once the kernel changes
have had more testing 'in the wild' themselves.