Allow default values to depend on other arguments (fixes #281, fixes #33)
This commit allows default values of xparse's D, G, O, R type arguments
to depend on the values of other arguments.  Since E-type arguments are
done quite differently (and are experimental) I did not figure out yet
how to make that work for them.  This is only available for
non-expandable commands.