Warn when optional arguments cannot be omitted (fixes #3)
With an argument specifications such as {gom} or {e{_^}or__}
the optional arguments cannot be completely omitted because the
following mandatory argument starts with the same token as an
optional one.  This is now detected but I thought it dangerous
to produce an error so I just put a warning.