Speed-up line-wrapping code
This commit speeds up the line-wrapping code by changing the
approach completely: instead of grabbing words one by one and
checking if they fit, we now grab a whole line and chop off
the last word if it does not fit.  Unfortunately, the speed-up
is smaller than I hoped.  None of the output is changed.