l3build: luaotfload normalisation, and diff printing
1. (luc: path/to/font.luc) -> (luc: [..]/font.luc)
2. when -H is used, print the diff directly.

The idea of #2 is that you're asking it to stop when coming to a problem, so it kind of makes sense to display that problem immediately. More pragmatically, I needed that to debug problems in remove installations where only l3build output was able to be queried (Travis CI in this case).

Happy to switch this behaviour to be optional if people prefer.