Re-write l3alloc
This is part of preparation for some other format-mode changes but
also pulls l3alloc more into line with LaTeX2e.

Whilst in principle we have a free, in many ways it makes sense
to stick to the plain TeX numbering for tracking here. In particular,
later check-ins will deal with some issues related to LuaTeX where
access from the 'Lua side' does need an amount of insight into the
low-level setup.