Make mandatory argument of -NoValue- remain -NoValue- (fixes #349)
Previously that became an empty argument (if the command had any
default arguments) because the machinery for default arguments replaces
-NoValue- by the corresponding default indiscriminately.  I simply put
a default value of -NoValue- for mandatory arguments.

As a side-effect, I've optimized the code so that a function such that
all default values are -NoValue- does not use the machinery at all.