Switch l3regex to using arrays of fontdimen instead of skip/dimen
This likely slows down l3regex but it means that the only set of
registers we "abuse" are toks.  With a bit more work we can make
sure we only used unallocated toks, like in l3sort.  Once that is
done we can allow users to run code during the regex matching
process.