typeset in subdirectories
As already cared for with unpacking, typeset files should also appear
besides they were created from. This is more complicated here, as the
source files originate from multiple directories. Because of that, these
directories have to be passed down through to the runtool function,
which executes all the external programs for document compilation and
sets up the paths.