Add a stub l3image
The plan here is to work on image inclusion, which is largely
handled at the driver end. This new module will provide a place
for shared code that doesn't properly belong in the driver files
themselves. Eventually there will be some form of
\includegraphics-like interface too.

This work is at least in part to help with supporting LaTeX2e:
the drivers there are a bit of a mess at present and some space
to work out what is actually required (and how it works) would be
very welcome.