Rs-add "pagebox" support in l3image
As raised on GitHub for the 2e driver set, there are applications
which produce various pagebox specs. Support is not complex, so
is re-introduced here. Note that the "cropbox" option is the standard
but is always given explicitly for PDF images (as this avoids any
binary variation in what is selected).