Store full bbox data for images, and use this for dvipdfmx
In the dvipdfmx case, "pagebox" will go wrong unless there is also a
full "bbox" argument. That means we need to store the information at
least in that case. It therefore makes life easier if we simply cache
the bbox in all cases, noting that where (llx,lly) = (0,0) we can
avoid consuming two dimens. In some cases, for example pdfmode,
xdvipdfmx, images always start (0,0) so that is a built-in assumption.