Correct value of \g_file_current_name_str
This largely relies on the fact that LaTeX2e already tracks the
value in the preamble using \@pushfilename/\@popfilename. To grab the
'new' values we have to do a bit of cheating, relying on the fact that
\@pushfilename is always followed by

    \xdef\@currname{#1}%

Currently, there is no attempt to pick up from \input or similar:
that needs something like the filehooks package and is likely too
invasive at the LaTeX2e for us.