Tell l3doc that a deprecated l3box function is deprecated
The current situation for deprecated functions is not good, as they are
listed with removal dates in three places:

- in l3obsolete.txt,
- in their respective modules with the old implementation and
    some code producing a warning,
- in l3deprecation with code that the user can replace them with.

In principle it would be better if the removal date was computed automatically
from the date we decide to deprecate the commands but that is difficult because
l3doc needs to know and the functions themselves need to know.  In any case,
l3obsolete has to be updated by hand at the end of the day so we don't gain much.