Rephrase future tense to present tense in l3kernel
I may have gone a bit overboard.  Basically I used present
tense to describe what a function does at time of use, and
future for its longer-term effect.  For instance see the
description of \cs_new_protected:Npn, which now says roughly
"an error is raised if the cs already exists" and "the cs
will not expand in x-type expansion".