Allow for Unix/Win line ending differences in file tests
As we checkout with line endings normalised for the host platform,
the exact details of the file size and hash we see varies.

One could deal with the hash by doing a string read of the entire
thing then reporting the hash for those tokens, as the 'back end'
will then have normalised the endings. However, that takes longer
and doesn't really fit with being 'file' ops.