Add \debug_log_functions_on/off: and \debug_check_declarations_on/off:
These functions apply the eponymous options of expl3 locally.  They can
only be used if expl3 is loaded with the new option "enable-debug", or
with "check-declarations" or "log-functions".  This is the only way to
avoid impacting performance for people who do not care about such checking.
The commit also applies \__debug_patch:nnNw in a few more places.