Extend internal colour model [ci skip]
This is a purely documentation change at this stage: allow for more models.

The thinking here is that this will cover spot colours and mixing of such colours
(see colorspace package for examples). It will also cover tints of spot colours:
I have included a mandatory 'amount' of spot.

I am retaining spaces as separators: this is closer to the dvips/PostScript/PDF
approach than commas, and is arguably more readable for this rather special case.
We can always revisit that: currently this is purely internal.

The second phase of this change will take place in l3drivers: code is needed
there.